% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_solve.R
\name{binary_solve}
\alias{binary_solve}
\title{Find solution of a 1d increasing function}
\usage{
binary_solve(f, interval, eps = 0.001, verbose = FALSE)
}
\arguments{
\item{f}{Input function.}

\item{interval}{A vector of length 2. Interval that contains the zero.}

\item{eps}{Terminate when the size of interval is less than eps.}

\item{verbose}{Print progress if \code{TRUE}}
}
\description{
Binary search for zeros of a 1-d increasing function.
Values at the left and right side of the interval should be negative and positive respectively.
}
\examples{
\dontrun{
f <- function(t) t
binary_solve(f, c(-2, 2), eps = 0.001)
}
}
