% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootglm.R
\name{bootglm}
\alias{bootglm}
\title{Compute resized bootstrap MLE for a GLM multiple times}
\usage{
bootglm(X, beta, family, b_boot, verbose)
}
\arguments{
\item{X}{A covariate matrix of size n * p.}

\item{beta}{A vector of length n. Coefficients where the parametric bootstrap is applied.}

\item{family}{An object of class family. Besides the usual definition of glm family (see [stats::family]), 
\code{family} should include an additional element \code{simulate_fun} (see [get_simulate_fun]).}

\item{b_boot}{An integer of number of bootstrap samples.}

\item{verbose}{Print progress if \code{T}.}
}
\value{
mle_boot a matrix of size p * b_boot of the bootstrap MLE
Returns error if the bootstrap MLE does not exist more than 20% of times.
}
\description{
This function implements a parametric bootstrap (for a GLM). Unlike the standard parametric bootstrap,
where we use the MLE to generate new responses, here user can input a coefficient vector beta. 
The function calls the the function boot_fun, which implements the resized bootstrap once, and returns a vector of MLE.
}
