% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equation_binary.R
\name{equation_binary}
\alias{equation_binary}
\title{Setup a system of equations}
\usage{
equation_binary(
  rho_prime,
  f_prime1,
  f_prime0,
  kappa,
  gamma,
  beta0,
  intercept = TRUE
)
}
\arguments{
\item{rho_prime}{A function that computes the success probability \eqn{\mathrm{P}(Y=1 | X^\top \beta)},
here \eqn{\beta} is the coefficient. The default is logistic model.}

\item{f_prime1}{A function. Derivative of the loss function when \eqn{Y = 1}.
The default is the derivative of the negative log-likelihood of logistic
regression when \eqn{Y = 1}.}

\item{f_prime0}{A function. Derivative of the loss function when \eqn{Y = -1}.
The default is the derivative of the negative log-likelihood of logistic
regression when \eqn{Y = -1}.}

\item{kappa}{Numeric. The problem dimension \eqn{\kappa = p/n}.}

\item{gamma}{Numeric. Signal strength \eqn{\gamma = \sqrt(\mathrm{Var}(X^\top \beta))}.}

\item{beta0}{Numeric. Intercept.}

\item{intercept}{If \code{TRUE}, the glm contains an intercept term.
\code{intercept = TRUE} by default.}
}
\value{
A function that takes as input the parameters \eqn{(\alpha,\lambda,\sigma,b)}
    and returns a vector of length 4, which is the value of the four equations.
    When the model contains no intercept term (\code{intercept = FALSE}),
    returns a system of three equations. The special case when there is no signal (\code{gamma = 0})
    or intercept, returns a system of two equations.
}
\description{
This function sets up a system of four equations for binary regressions.
The solution to the system characterizes the asymptotic bias and variance
of the M-estimator, as well as the Hessian of the loss function (in case
of the MLE, the loss function is the negative log-likelihood).
}
\details{
Following is the formula of the four equations:
\deqn{
 \begin{dcases}
 \sigma^2\kappa^2 & =\E{\rho'(S_1)(\lambda\rho'(\mathrm{prox}_{\lambda\rho}(-S_2)))^2+\rho'(-S_1)(\lambda\rho'(\mathrm{prox}_{\lambda\rho}(S_2)))^2}\\
 \sigma \sqrt{\kappa}(1-\kappa) & =\E{\rho'(S_1)Z_2\mathrm{prox}_{\lambda\rho}(\lambda+S_2) + \rho'(-S_1)Z_2\mathrm{prox}_{\lambda\rho}(S_2)}\\
 \gamma_0 \alpha & = \E{\rho'(S_1)Z_1\mathrm{prox}_{\lambda\rho}(\lambda+S_2) + \rho'(-S_1)Z_1\mathrm{prox}_{\lambda\rho}(S_2)},
 0 & = \E{-\rho'(S_1)\rho'(\mathrm{prox}_{\lambda\rho}(-S_2)) + \rho'(-S_1)\rho'(\mathrm{prox}_{\lambda\rho}(S_2))}.
 \end{dcases}
}
where \eqn{(Z_1, Z_2)\sim\mathcal{N}(0, I_2)} and
\deqn{
  S_1 = \gamma_0 Z_1 + \beta_0 ,\quad S_2 = \alpha \gamma_0 Z_1 + \sigma\sqrt{\kappa} Z_2 + b_0,
}
When the variables does not have an intercept term, then \eqn{b_0 = 0}.
If the model does not have an intercept, then \eqn{\beta_0 = 0}.
}
\references{
\emph{A modern maximum-likelihood theory for high-dimensional logistic regression},
Pragya Sur and Emmanuel J. Candes,
Proceedings of the National Academy of Sciences Jul 2019, 116 (29) 14516-14525
}
\keyword{internal}
