% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_eta.R
\name{estimate_eta}
\alias{estimate_eta}
\title{Estimate eta from MLE coef.}
\usage{
estimate_eta(X, y, beta_hat, family)
}
\arguments{
\item{X}{A covariate matrix of size n*p.}

\item{y}{A vector of responses of length n.}

\item{beta_hat}{The MLE vector of length p.}

\item{family}{A GLM family, with family and link. See also [compute_deriv()].}
}
\value{
A numeric value of the estimated \eqn{\hat{\eta}}.
}
\description{
Use an extension of the SLOE estimator to estimate \eqn{\eta = \var(X_{\mathrm{new}}^\top \hat{\beta})^{1/2}},
where \eqn{X_{\mathrm{new}}} is a new obs. and \eqn{\hat{\beta}} is the MLE.
}
\details{
Let \eqn{f_y(t)} be the negative log-likelihood when the response is \eqn{y} and linear predictor is \eqn{t}.
Define \eqn{w_i = x_i^\top H^{-1}x_i} and \eqn{t_i = \ x_i^\top \hat{\beta}}, where \eqn{x_i} is the i-th obs., and
\eqn{H} is the Hessian of the negative log-likelihood evaluated at the \eqn{\hat{\beta}}. Let
\deqn{
S_i = x_i^\top \hat{\beta} + q_i f'_{y_i}(t_i),
}
where 
\deqn{
q_i = \frac{w_i}{1-w_i f''_{y_i}(t_i)}.
}
Then, the SLOE estimator is defined to be 
\deqn{
\hat{\eta}^2  = \frac{1}{n}\sum_{i=1}^n S_i^2 - \left(\frac{1}{n}\sum_{i=1}^n S_i\right)^2.
}
}
