% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_gamma.R
\name{estimate_gamma}
\alias{estimate_gamma}
\title{Estimate the signal strength}
\usage{
estimate_gamma(s_seq, eta_hat, eta_obs, sd_obs, verbose = T, filename = NA)
}
\arguments{
\item{s_seq}{A sequence of shrinkage factors s.}

\item{eta_hat}{A matrix. The number of rows is equal to the length of s_seq, the number of columns is equal to the number of parametric bootstrap samples at each s.}

\item{eta_obs}{\eqn{\hat{\eta}} computed using observed samples.}

\item{sd_obs}{Observed standard deviation of the linear predictors evaluated at \eqn{\hat{\beta}}, i.e., \eqn{sd(x_i^\top \hat{\beta})}.}

\item{verbose}{Plot \eqn{\gamma} versus shrinkage factors \eqn{s} if \code{TRUE}.}

\item{filename}{If a file name is provided, then save the plot of \eqn{\hat{\eta}(s)} versus \eqn{\gamma} to \code{filename}.}
}
\value{
\describe{
\item{s_hat}{A numeric value of the estimated shrinkage factor \eqn{s} that satisfies
\eqn{\hat{s} \mathrm{sd}(X^\top \hat{\beta}) = \hat{\gamma}.}
}
\item{gamma_hat}{A numeric value of the estimated signal strength.}
}
}
\description{
Estimating gamma using the SLOE estimator and parametric bootstrap
}
\details{
We estimate \eqn{\gamma} by the standard deviation of \eqn{sd(x_i^\top \beta(s_{\star}))}, 
  where \eqn{\beta(s_\star) = s_\star \cdot \hat{\beta}}. We use the following relationship: if 
  \eqn{sd(x_i^\top \beta(s_{\star}))\approx \gamma}, then \eqn{\hat{\eta}(s_\star) \approx \hat{\eta}}.
  In this equation, \eqn{\hat{\eta}} is the estimated \eqn{\hat{\eta}} from the observations,
  \eqn{\eta(s)} is \eqn{\eta} when the model coefficient is \eqn{\beta(s) = s\cdot\hat{\beta}}. We 
  estimate \eqn{\eta(s)} by parametric bootstrap, fixing the covariates at the observed values and setting the model 
  coefficients as \eqn{\beta(s)} (see [estimate_variance] on how to estimate \eqn{\eta(s)}). 
  
We pick a sequence of shrinkage factors \eqn{s}, and then compute \eqn{\hat{\eta}(s)} for each of them.
  Then, we fit a LOESS curve of \eqn{\eta(s)} as a function of \eqn{s} using the sequence of \eqn{s} and 
  the estimated \eqn{\hat{\eta}(s)} at each bootstrap sample. Finally, we choose the shrinkage factor \eqn{s_\star} on the 
  curve such that the fitted value is equal to the observed \eqn{\hat{\eta}}. 
  
The estimated \eqn{\hat{\gamma}} is \eqn{sd(x_i^\top \beta(s_{\star}))}.
}
