% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_variance.R
\name{estimate_variance}
\alias{estimate_variance}
\title{Estimate std.dev. of the linear predictor evaluated at the MLE}
\usage{
estimate_variance(x, y, beta, family, b_var)
}
\arguments{
\item{y}{A vector of responses of length n}

\item{beta}{A vector of true model coef.}

\item{family}{A GLM family, with family and link. See also [compute_deriv()].}

\item{b_var}{Numeric. Number of bootstrap replicates used to estimate \eqn{sd(X^\top \hat{\beta})}.}
}
\value{
A vector of length \code{b_var} of \eqn{\hat{\eta}} in each bootstrap sample. 
  if the GLM function reports an error in more than 50% of times, then return -1.
}
\description{
Estimates \eqn{\eta = sd(X_\mathrm{new}^\top \hat{\beta})} where \eqn{X_{\mathrm{new}}} is a new obs. and \eqn{\hat{\beta}} is the MLE
when the true coefficient is \eqn{\beta}
}
\details{
Generate \code{b_var} parametric bootstrap samples by sampling new responses \eqn{y} at the
observed covariates and when the model coef. is \code{\beta}. 
Use SLOE to estimate \eqn{\eta} in each bootstrap sample.
}
