% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_simulate_fun.R
\name{get_simulate_fun}
\alias{get_simulate_fun}
\title{Simulate response given linear predictors}
\usage{
get_simulate_fun(family)
}
\arguments{
\item{family}{A family object, it contains a GLM family \code{family} and an inverse link function \code{linkinv}.
Currently supports binomial and Poisson families.}
}
\value{
simulate_fun a function which takes as input a covariate matrix X and a coefficient vector beta.
and returns a vector containing simulated responses given X andbeta.
}
\description{
Simulate response given linear predictors
}
\examples{
family <-  binomial(link = "logit")
fun <- get_simulate_fun(family)
y <- fun(X = matrix(rnorm(100, 0, 1), 10, 10), beta = rnorm(10))
}
