% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_glmadj.R
\name{glmadj}
\alias{glmadj}
\alias{print.glmadj}
\alias{summary.glmadj}
\alias{print.summary.glmadj}
\alias{predict.glmadj}
\title{Methods for adjusted glm objects}
\usage{
\method{print}{glmadj}(x, ...)

\method{summary}{glmadj}(object, ...)

\method{print}{summary.glmadj}(x, ...)

\method{predict}{glmadj}(object, newdata = NULL, ...)
}
\arguments{
\item{x}{A \code{glmadj} object, created from \code{\link{adjust_glm}}.}

\item{...}{Additional parameters}

\item{object}{A \code{glmadj} object, created from \code{\link{adjust_glm}}.}

\item{newdata}{A data frame created with \link[stats]{model.frame}.}
}
\value{
\code{\link{summary.glmadj}} returns a summary table with four columns
\itemize{
\item ajusted_mle \eqn{\hat{\beta}_j^\mathrm{Adj} = \hat{\beta}_j^\mathrm{MLE} / \alpha_\star}.
\item adjusted_std. Standard error of the adjusted MLE \eqn{\hat{\sigma}_j / \alpha_\star = \sigma_\star / \alpha_\star \tau_j}.
\item t.value \eqn{t_j = \hat{\beta}_j^\mathrm{MLE} / \hat{\sigma}_j}. When \eqn{\beta_j = 0}, \eqn{t_j} is approximately
    a standard Gaussian as \eqn{n,p \to\infty}.
\item p.value 2-sided p-value using t.value to test whether \eqn{\beta_j = 0}.
}
}
\description{
Availabel methods for \code{glmadj} objects: print coefficients, summary,
prediction
}
\seealso{
\code{\link{adjust_glm}}
}
