% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate2_normal.R
\name{integrate2_normal}
\alias{integrate2_normal}
\title{2-dimensional Gaussian Expectation}
\usage{
integrate2_normal(f, ...)
}
\arguments{
\item{f}{Function to integrate over. Input to
\code{f} should be a length 2 vector.}

\item{...}{Additional arguments to \link[cubature]{hcubature}}
}
\description{
Expected value of a function of two independent
standard Gaussian variables
}
\examples{
\dontrun{
# Second moment of a Gaussian
# tol = 0.0001, true value = 2, evaluated = 2.000001
f <- function(x) x[1]^2 + x[2]^2
integrate2_normal(f, tol = 1e-4)
}
}
