% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_anova.R
\name{lrt_glm}
\alias{lrt_glm}
\title{LRT for high-dimensional glm}
\usage{
lrt_glm(object, param)
}
\arguments{
\item{object}{A list with at least two elements from the class \link[stats]{glm}.
The function does not check whether the models are nested.}

\item{param}{A named vector with elements \code{sigma_s} and \code{lambda_s} which
can be outputs from the \code{adjust_glm} function.}
}
\value{
A list with two elements
\describe{
\item{models}{Model formula}
\item{anova.tab}{Significance tables for the LRT}
}
}
\description{
This is the generic function to compute the likelihood ratio statistics
and p-valus for binary regression models.
}
\details{
Under the null hypothesis:
\deqn{
H_0:\quad \beta_1 = \ldots = \beta_k = 0,
}
two times the likelihood ratio statistics \eqn{\Lambda} is asymptotically from a
re-scaled chi-squared distribution with \eqn{k} degrees of freedom
\deqn{
2 \Lambda \stackrel{d}{\longrightarrow} \frac{\sigma_\star^2}{\lambda_\star}\chi^2_k
}
where \eqn{(\sigma_\star, \lambda_\star)} can be computed using \code{find_param} function.
}
\examples{
\dontrun{
n <- 1000L
p <- 300L
X <- matrix(rnorm(n*p, 0, 1), n, p) / sqrt(p)
beta <- rep(c(0, 1), each = p / 2)
Y <- rbinom(n, 1, 1 / (1 + exp(- X \%*\% beta)))
f1 <- glm(Y ~ X + 0, family = binomial, x = TRUE, y = TRUE)
f2 <- glm(Y ~ X[ , -1] + 0, family = binomial, x = TRUE, y = TRUE)
adjusted_fit <- adjust_glm(f1)
lrt_glm(list(f1, f2), adjusted_fit$param)
}
}
\references{
\emph{The likelihood ratio test in high-dimensional logistic regression is asymptotically a rescaled Chi-square},
Pragya Sur, Yuxin Chen and Emmanuel Candes, Probab. Theory Relat. Fields 175, 487–558 (2019).
}
