% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probe_frontier.R
\name{probe_frontier}
\alias{probe_frontier}
\title{Estimate the problem dimension where two classes become linearly separable}
\usage{
probe_frontier(X, Y, B = 10, eps = 0.001, verbose = FALSE)
}
\arguments{
\item{X}{Covariate matrix. Each row in \code{X} is one observation.}

\item{Y}{Response vector of \eqn{+1} and \eqn{-1} representing
the two classes. \code{Y} has the same length as the number of rows
in \code{X}.}

\item{B}{Numeric. How many subsamples should I generate
for each sample size?}

\item{eps}{Numeric. Minimum window size.
Terminate when the search interval is smaller than \code{eps}}

\item{verbose}{Print prgress if \code{TRUE}.}
}
\value{
Numeric. Estimated kappa_hat.
}
\description{
This function estimates the sample size \eqn{n_s}, or equivalently problem dimension
\eqn{\kappa_s = p/n_s}, that two classes from the data becomes separable. To locate \eqn{\kappa_s},
we bisect the interval \eqn{[p/n, 0.5]}, until the window
size is smaller than \code{eps}. For each sample size \code{nn}, it generates
\code{B} subsamples of size \code{nn}, and estimate the separable probability
\eqn{\hat{\pi}} with the proportion of separable subsamples.
Finally we fit a logistic regression using \eqn{\hat{\pi}} as response
and \eqn{\kappa = p/nn} as covariate to determine the \eqn{\hat{\kappa}}
where separable probability is 0.5.
}
\examples{
# Y is independent of X, kappa_s is approximately 0.5
n <- 1000; p <- 200
X <- matrix(rnorm(n*p, 0, 1), n, p)
Y <- 2 * rbinom(n, 1, 0.5) - 1
probe_frontier(X, Y, verbose = TRUE)
}
