% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_link.R
\name{process_link}
\alias{process_link}
\title{Extract likelihood from link function}
\usage{
process_link(glm_family)
}
\arguments{
\item{glm_family}{A glm \code{familly} object for \code{binomial} family.
It should have an element called \code{link}.}
}
\value{
A list with three elements
\describe{
\item{rho_prime}{A function that computes the success probabily \eqn{\rho'(t) = \mathrm{P}(Y=1|X^\top \beta = t)}.}
\item{f_prime1}{Derivative of the negative log-likelihood when \eqn{Y=1}}
\item{f_prime0}{Derivative of the negative log-likelihood when \eqn{Y=0}}
}
}
\description{
This function returns success probability and derivatives of the log-likelihood
for commonly used link function in the GLM binary family.
}
\examples{
\dontrun{
process_link(binomial(link = "cloglog"))
}
}
\keyword{internal}
