% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_model.R
\name{rho_prime_logistic}
\alias{rho_prime_logistic}
\alias{f_prime1_logistic}
\alias{f_prime0_logistic}
\title{Logistic loss function}
\usage{
rho_prime_logistic(x)

f_prime1_logistic(x)

f_prime0_logistic(x)
}
\arguments{
\item{x}{Numeric. Input variable.}
}
\description{
Success probability and derivatives of the
negative log-likelihood of a logistic model
}
\details{
\code{rho_prime_logistic} computes the successs probability of
a logistic model \eqn{\rho'(t) = 1/(1+e^{-t})}.

\code{f_prime1_logistic} and \code{f_prime0_logistic} are
the derivative of the negative log-likelihood when \eqn{Y=1} and \eqn{Y=-1}.
}
