% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal_strength.R
\name{signal_strength}
\alias{signal_strength}
\title{Estimate intercept and signal strength}
\usage{
signal_strength(
  rho_prime = rho_prime_logistic,
  kappa_hat,
  intercept = FALSE,
  p0 = NA,
  verbose = FALSE,
  tol = 1e-04
)
}
\arguments{
\item{rho_prime}{A function that computes the success probability \eqn{\rho'(t) = \mathrm{P}(Y=1 | X^\top \beta = t)},
here \eqn{\beta} is the coefficient. The default is logistic model.}

\item{kappa_hat}{Numeric. Estimated dimension where the data becomes linearly separable}

\item{intercept}{Logical \code{TRUE} if the model contains an intercept.}

\item{p0}{Numeric. Proportion of outcomes \eqn{Y=1}.}

\item{verbose}{Should progress be printed? If \code{TRUE}, prints progress at each step.}

\item{tol}{Numeric. Tolerance to be used in \code{fsolve} function.}
}
\value{
If the model does not contain an intercept, returns estimated \code{gamma_hat}.
    Otherwise, returns a list with two components
\describe{
  \item{gamma_hat}{Estimated signal strength.}
  \item{b_hat}{Estimated intercept.}
}
}
\description{
\code{signal_strength} estimates \eqn{(\beta_0, \gamma)} using the estimated \eqn{\kappa_s}
and observed proportion of successes.
}
\details{
Assume that \eqn{Y} depends on \eqn{X} as
\deqn{
\mathrm{P}(Y=1\,|\,X) = \rho'(X^\top \beta + \beta_0),
}
and let the signal strength be \eqn{\gamma = \mathrm{Var}(X^\top \beta)^{1/2}}.
The pair \eqn{(\beta_0, \gamma)} satisfies
\itemize{
\item They are on the phase transition curve \eqn{\kappa(\beta_0, \gamma)}.
    \deqn{
    \hat{\kappa} \approx \kappa(\beta_0, \gamma)
     }
\item The observed proportion of \eqn{Y=1} should be close to the expected proportion.
    \deqn{
    p_0 \approx \prob(Y = 1\,|\, \beta_0, \gamma) = \mathrm{E}{\mathrm{Ber}(\rho'(\beta_0 + \gamma Z)}
    }
   where \eqn{Z} is a standard normal variable.
}
We solve the above system of two equations to obtain an estimate of \eqn{(\beta_0, \gamma)}
}
\examples{
\dontrun{
# no signal case
# should return 0, returns 0.0127
signal <- signal_strength(kappa_hat = 0.5, intercept = FALSE)
signal$gamma_hat
}
}
