% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h_eq.R
\name{solve_kappa}
\alias{solve_kappa}
\alias{solve_beta}
\alias{solve_gamma}
\title{Compute the phase transition curve}
\usage{
solve_kappa(rho_prime, beta0, gamma0)

solve_beta(rho_prime, kappa, gamma0, verbose = FALSE)

solve_gamma(rho_prime, kappa, beta0, verbose = FALSE)
}
\arguments{
\item{rho_prime}{Function. Success probability \eqn{\rho(t) = \mathrm{P}(Y=1\,|\, X^\top \beta = t)}}

\item{beta0}{Numeric. Intercept value.}

\item{gamma0}{Numeric. Signal strength.}

\item{kappa}{Numeric. Problem dimension on the phase transition curve.}

\item{verbose}{Print progress if \code{TRUE}.}
}
\value{
Numeric. Problem dimension \eqn{\kappa} (\eqn{\beta} or \eqn{\gamma}) on the phase transition curve.
}
\description{
\code{solve_kappa} computes the problem dimension \eqn{\kappa} where
the phase transition occurs in binary regression, given \eqn{\beta} and \eqn{\gamma_0}. \code{solve_beta} and \code{solve_gamma}
computes \eqn{\beta_0} and \eqn{\gamma_0} on the phase transition curve
given the other one and \eqn{\kappa}.
}
\details{
When covariates are multivariate Gaussian, the phase transition dimension
can be characterized as following.
\deqn{
&\kappa > h_{\mathrm{MLE}}(\beta_0, \gamma_0) \implies \lim_{n,p\to\infty} \mathrm{P}(\text{MLE exists}) = 0\\
&\kappa < h_{\mathrm{MLE}}(\beta_0, \gamma_0) \implies \lim_{n,p\to\infty} \mathrm{P}(\text{MLE exists}) = 1.
}
The function \eqn{h} is defined to be
\deqn{
h_{\mathrm{MLE}}(\beta_0, \gamma_0) = \min_{t_0, t_1 \in \mathbb{R}} \mathbb{E}\left[(t_0 Y + t_1 V - Z)_+^2 \right],
}
where \eqn{X\sim\mathcal{N}(0,1)} and \eqn{\mathrm{P}(Y=1|X) = 1- \mathrm{P}(Y=-1|X) = \rho'(\beta_0 + \gamma_0 X) }.
\eqn{Z\sim\mathcal{N}(0,1)} and is independent of \eqn{X,Y}. The phase transition
curve is thus \eqn{\kappa(\beta_0, \gamma_0)}. It also depends on the
success probability \eqn{\rho'}.
}
\examples{
\dontrun{
# when Y is independent of X, should return 0.5 for logistic model
# should return 0.5
rho_prime_logistic <- function(t) 1 / (1 + exp(-t))
solve_kappa(rho_prime_logistic, 0, 0)
}
}
\references{
\emph{The phase transition for the existence of the maximum likelihood estimate in high-dimensional logistic regression}
Emmanuel J. Candes and Pragya Sur, Ann. Statist., Volume 48, Number 1 (2020), 27-42.
}
