% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_fun.R
\name{boot_fun}
\alias{boot_fun}
\title{Compute one bootstrap for a GLM}
\usage{
boot_fun(X, beta, family, sloe = F)
}
\arguments{
\item{X}{A n*p covariate matrix (each row represents one obs.).}

\item{beta}{Model coefficient.}

\item{family}{GLM family. Besides the usual definition of glm family (see [stats::family]), 
\code{family} should include an additional element \code{simulate_fun} (see [get_simulate_fun]).}

\item{sloe}{If True, use SLOE to estimate \eqn{\eta = \var(X_{\mathrm{new}^\top \hat{\beta}})^{1/2}} 
(see the function \code{estimate_eta}).}
}
\value{
\describe{
\item{coef}{A length p vector of MLE coefficients.}
\item{eta_hat}{If \code{sloe = T}, returns the estimated \eqn{\hat{\eta}}.}
}
}
\description{
Generates one parametric bootstrap sample given GLM family and model coefficients.
}
\details{
This model does not fit an intercept. As a result, if you would like to fit an intercept,
please include a column of 1 in the matrix X.
}
