% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_simulate_fun.R
\name{get_simulate_fun}
\alias{get_simulate_fun}
\title{Simulate response given linear predictors}
\usage{
get_simulate_fun(family)
}
\arguments{
\item{family}{A family object, it contains a GLM family \code{family} and an inverse link function \code{linkinv}}
}
\value{
simulate_fun a function which takes a vector of linear predictors as inputs
and returns a vector of simulated responses given the linear predictors.
}
\description{
Simulate response given linear predictors
}
\examples{
family <-  binomial(link = "logit")
fun <- get_simulate_fun(family)
y <- fun(rnorm(10))
}
