% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_glmadj.R
\name{glmadj}
\alias{glmadj}
\alias{print.glmadj}
\alias{summary.glmadj}
\alias{print.summary.glmadj}
\alias{predict.glmadj}
\title{Methods for adjusted glm objects}
\usage{
\method{print}{glmadj}(x, ...)

\method{summary}{glmadj}(object, ...)

\method{print}{summary.glmadj}(x, ...)

\method{predict}{glmadj}(object, newdata = NULL, ...)
}
\arguments{
\item{x}{A \code{glmadj} object, created from \code{\link{adjust_glm}}.}

\item{...}{Additional parameters}

\item{object}{A \code{glmadj} object, created from \code{\link{adjust_glm}}.}

\item{newdata}{A data frame created with \link[stats]{model.frame}.}
}
\description{
Availabel methods for \code{glmadj} objects: print coefficients, summary,
prediction
}
\seealso{
\code{\link{adjust_glm}}
}
