% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal_strength.R
\name{signal_strength}
\alias{signal_strength}
\title{Estimate problem parameters}
\usage{
signal_strength(
  rho_prime = rho_prime_logistic,
  f_prime1 = f_prime1_logistic,
  f_prime0 = f_prime0_logistic,
  kappa_hat,
  kappa = NULL,
  intercept,
  b_hat = NULL,
  verbose = FALSE,
  eps = 0.1
)
}
\arguments{
\item{rho_prime}{A function that computes the success probability \eqn{\rho'(t) = \mathrm{P}(Y=1 | X^\top \beta = t)},
here \eqn{\beta} is the coefficient. The default is logistic model.}

\item{f_prime1}{A function. Derivative of the loss function when \eqn{Y = 1}.
The default is the derivative of the negative log-likelihood of logistic
regression when \eqn{Y = 1}.}

\item{f_prime0}{A function. Derivative of the loss function when \eqn{Y = -1}.
The default is the derivative of the negative log-likelihood of logistic
regression when \eqn{Y = -1}.}

\item{kappa_hat}{Numeric. Estimated dimension where the data becomes linearly separable}

\item{kappa}{Numeric. Problem dimension \eqn{\kappa = p/n}.}

\item{intercept}{\code{TRUE} if the model contains an intercept.}

\item{b_hat}{Numeric. MLE of the intercept.}

\item{verbose}{Should progress be printed? If \code{TRUE}, prints progress at each step.}

\item{eps}{Numeric. Terminate the algorithm if the search interval is smaller than \code{eps}.}
}
\value{
A list of two components
\describe{
  \item{gamma_hat}{Estimated signal strength}
  \item{beta_hat}{Estimated intercept. \code{NULL} if the model does not have an intercept term}
}
}
\description{
This function estimates \eqn{(\beta_0, \gamma_0)} using the estimated \eqn{\kappa_s}
and intercept \eqn{b}.
}
\details{
Assume that \eqn{Y} depends on \eqn{X} as
\deqn{
\mathrm{P}(Y=1\,|\,X) = \rho'(X^\top \beta + \beta_0),
}
and let the signal strength be \eqn{\gamma = \mathrm{Var}(X^\top \beta)^{1/2}}.
The pair \eqn{(\beta_0, \gamma_0)} satisfies
\itemize{
\item They are on the phase transition curve \eqn{\kappa(\beta_0, \gamma_0)}.
    We estimate the dimension \eqn{\kappa(\beta_0, \gamma_0)} in \code{\link{probe_frontier}}.
\item The set of parameters \eqn{(\kappa, \beta_0, \gamma_0)} determines the asymptotic
    an estimated intercept \eqn{b}, which should be close to the MLE \code{b_hat}.
}
We use bisection to search for \eqn{\beta_0}, for each \eqn{\beta_0} we compute \eqn{\gamma_0}
on the phase transition curve, and compare the corresponding \eqn{b} with the
MLE \code{b_hat}. The algorithm terminates when search window is smaller than \code{eps}.
}
\examples{
\dontrun{
# no signal case
# should return 0, returns 0.0127
signal <- signal_strength(kappa_hat = 0.5, intercept = FALSE)
signal$gamma_hat
signal$b_hat
}
}
